Refactor policy DB module
